+++ app/darkspinner/icon.go
@@ -4,3 +4,6 @@ import _ "embed"
 
 //go:embed build/appicon.png
 var darkSpinnerIcon []byte
+
+//go:embed build/windows/icon.ico
+var darkSpinnerTrayIcon []byte